requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3

# HTTP Client
httpx==0.25.0

# Database
sqlalchemy==2.0.22
asyncpg==0.28.0
alembic==1.12.0
psycopg2-binary==2.9.9

# Environment & Config
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging & Monitoring
structlog==23.2.0

# Queue & Task Processing
celery==5.3.4
redis==5.0.1
flower==2.0.1
